Integrated NelderMead mixin into VennUtils.
